A reorganization of the input_data class. Added new files for data_container.
