+++ src/input_data.cc
@@ -9,43 +9,6 @@
 
 
 ////////////////////////////////////////
-// data_container
-////////////////////////////////////////
-
-data_container::data_container( string _parameter_name, int _number_of_options, int _number_of_fields,
-                                string *_data_fields, int *_interpolate_fields ):
-                                parameter_name(_parameter_name),
-                                number_of_options(_number_of_options),
-                                number_of_fields(_number_of_fields)
-{
-  // copy data from arrays to vectors
-  data_fields        = vector<string> (_data_fields, _data_fields + _number_of_fields);
-  interpolate_fields = vector<int> (_interpolate_fields, _interpolate_fields + _number_of_fields);
-}
-
-////////////////////////////////////////
-
-data_container::~data_container()
-{
-}
-
-////////////////////////////////////////
-
-void data_container::create_data_vector()
-{
-  vector<double> empty_array(number_of_fields);                    // create a placeholder for data
-  for(int i=0;i<number_of_fields;i++) empty_array[i] = NAN;        // fill it with NANs
-
-  data.reserve(number_of_points);                                  // reserve proper amount of memory
-  for(int i=0;i<number_of_points;i++) data.push_back(empty_array); // fill the vector with empty data
-}
-
-
-////////////////////////////////////////
-// input_data
-////////////////////////////////////////
-
-////////////////////////////////////////
 // Public methods
 ////////////////////////////////////////
 
@@ -73,7 +36,7 @@ void input_data::initialize()
 
 void input_data::load_data()
 {
-  read_data( *cascade_xsec_NN );
+  cascade_xsec_NN->read_data();
 }
 
 
@@ -104,114 +67,32 @@ void input_data::initialize_input_path()
 
 void input_data::initialize_data_containers()
 {
-  // Provide the name of the parameter that governs the data, the number of options,
+  // Provide  the parameter that governs the data,
   // then the number of different fields in the file, their names and the method of interpolation for each.
-  int cascade_xsec_NN_number_of_fields     = 10;
-  string cascade_xsec_NN_data_fields[]     = {"energy",
-                                              "xsec_ii", "xsec_ij",
-                                              "inel_ii", "inel_ij", "inel_1pi",
-                                              "angle_A_ii", "angle_A_ij", "angle_B_ii","angle_B_ij"};
-  int cascade_xsec_NN_interpolate_fields[] = {0, 0, 0, 0, 0, 0, 0, 0, 0, 0};
-  cascade_xsec_NN = new data_container( "kaskada_xsec_NN", 2, cascade_xsec_NN_number_of_fields,
-                               cascade_xsec_NN_data_fields, cascade_xsec_NN_interpolate_fields );
 
-  if ( par.kaskada_xsec_NN < cascade_xsec_NN->number_of_options )  // if the parameter is ok
+    string cascade_xsec_NN_file_name         = generate_file_name( "kaskada_xsec_NN", par.kaskada_xsec_NN );
+    int cascade_xsec_NN_number_of_fields     = 10;
+    string cascade_xsec_NN_data_fields[]     = {"energy",
+                                                "xsec_ii", "xsec_ij",
+                                                "inel_ii", "inel_ij", "inel_1pi",
+                                                "angle_A_ii", "angle_A_ij", "angle_B_ii","angle_B_ij"};
+    int cascade_xsec_NN_interpolate_fields[] = {0, 0, 0, 0, 0, 0, 0, 0, 0, 0};
+    cascade_xsec_NN = new data_container( cascade_xsec_NN_file_name, cascade_xsec_NN_number_of_fields,
+                                          cascade_xsec_NN_data_fields, cascade_xsec_NN_interpolate_fields );
+}
+
+////////////////////////////////////////
+
+string input_data::generate_file_name( string parameter_name, int parameter_option )
+{
+  if ( parameter_option < 2 )  // ckeck if the parameter is ok, it should be read from params, hardcoded for now
   {
-    generate_file_name( *cascade_xsec_NN, par.kaskada_xsec_NN );
+    stringstream name_sstream;
+    name_sstream << input_path << parameter_name << "_" << parameter_option << ".dat"; // path + name + extension
+    return name_sstream.str();
   }
   else
   {
     throw "input_data error: Invalid parameter.";
-  }
-}
-
-////////////////////////////////////////
-
-void input_data::generate_file_name( data_container &container, int option )
-{
-  stringstream name_sstream;
-  name_sstream << input_path << container.parameter_name << "_" << option << ".dat"; // path + name + extension
-  container.file_name = name_sstream.str();
-}
-
-////////////////////////////////////////
-
-void input_data::read_data( data_container &container )
-{
-  ifstream file_ifstream;
-  file_ifstream.open( container.file_name.c_str() );        // open the file
-
-  if( file_ifstream.is_open() )
-  {
-    string file_line;
-
-    // first check the number of data points
-    container.number_of_points = 0;                         // make sure its zero
-
-    while( getline ( file_ifstream, file_line ) )
-    {
-      if( file_line[0] == '-' )
-      {
-        container.number_of_points++;
-      }
-    }
-    container.create_data_vector();                         // reserve proper amount of memory
-                                                            // and fill the vector with empty data
-
-
-    file_ifstream.clear();                                  // go back to the start of the file
-    file_ifstream.seekg(0, ios::beg);
-
-
-    // then iterate through points
-    int point = -1;                                         // which data point
-    size_t char_position;                                   // position of a given char
-    string field;                                           // which field
-    double value;                                           // what is the data
-
-    while( getline ( file_ifstream, file_line ) )
-    {
-      if( file_line[0] == '#' )                             // a comment starts with #
-      {
-        continue;
-      }
-      if( file_line[0] == '-' )                             // new point starts after -
-      {
-        point++;
-        continue;
-      }
-      char_position = file_line.find( ':' );                // ":" means data
-      if( char_position != string::npos )                   // we have something to take care of
-      {
-        field = file_line.substr(0, char_position);         // erase everything up to ":"
-        field.erase(0, field.find_first_not_of(" \n\r\t") );// trim from left
-
-        value = stod( file_line.substr(char_position+1) );  // everything after ":", convert to double
-
-        for( int i=0; i<container.number_of_fields; i++ )   // determine the row and fill
-        {
-          if( container.data_fields[i] == field )
-          {
-            container.data[point][i] = value;
-            break;
-          }
-
-          if( i == container.number_of_fields-1 )
-          {
-            throw "input_data error: Invalid syntax.";
-          }
-        }
-      }
-    }
-    for(int i=0;i<container.data.size();i++){
-    for(int j=0;j<container.data[i].size();j++)
-      cout << container.data[i][j] << "\t";
-    cout << "\n";}
-
-    file_ifstream.close();                                  // close the file
-  }
-  else
-  {
-    throw "input_data error: Could not open the data file.";
   }
 }
